Extract code generating academic year dropdown into its own function
The Faculty and Courses tab both have academic year dropdowns in which the options are in the format "Fall {academicYear - 1}, Spring {academicYear}." To avoid duplicating the code that generates these options, the function that is generating these options was extracted to its own file so that the function can be used in both the Courses and Faculty page.
